Produce human readable reference number
In order to achieve this we have had to switch the primary key for the
database from a UUID to an incrementing integer, as this is not a change
that is easy to do with migrations we have recreated the migrations
which will require us to drop the tables in the development database.

We are generating the reference id using a library called Hashids. When
given a salt, Hashids produces a reversible hash that will be unique for
any number passed to it. Therefore we are not persisting the reference
in our database although it will be persisted in MCM and sent to the
user.